rename the http client to use for both instance and devices
